Day ten. Part two.
My approach for the real data has a bug, but not a fatal one. It's supposed to
return two sets, one outside and one inside. For the test data that worked fine.
For my real data the inside set filled all the spare squares (do to some, at
this point, unknown bug). The outside set worked ok, so that let me grab the
inside set (and the answer). I will, possibly, get to the bottom of the cause of
that problem ... or not.

This bug became obvious when I wrote the graphical renderer, which as fun and
cool.